refactor: move each class appropriate package
