+++ src/test/java/baseball/model/referee/RefereeTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import baseball.model.pitcher.Pitcher;
-import baseball.view.ScoreStatus;
+import baseball.status.ScoreStatus;
 import java.lang.reflect.Field;
 import java.util.Map;
 import java.util.stream.Stream;
